+++ index.html
@@ -193,27 +193,27 @@
 
   <!-- Карточка объявления -->
   <template id="card">
-<article class="popup">
-  <img src="img/avatars/user01.png" class="popup__avatar" width="70" height="70" alt="Аватар пользователя">
-  <h3 class="popup__title">Уютное гнездышко для молодоженов</h3>
-  <p class="popup__text popup__text--address">102-0082 Tōkyō-to, Chiyoda-ku, Ichibanchō, 14−3</p>
-  <p class="popup__text popup__text--price">5200 <span>₽/ночь</span></p>
-  <h4 class="popup__type">Квартира</h4>
-  <p class="popup__text popup__text--capacity">2 комнаты для 3 гостей</p>
-  <p class="popup__text popup__text--time">Заезд после 14:00, выезд до 10:00</p>
-  <ul class="popup__features">
-    <li class="popup__feature popup__feature--wifi"></li>
-    <li class="popup__feature popup__feature--dishwasher"></li>
-    <li class="popup__feature popup__feature--parking"></li>
-    <li class="popup__feature popup__feature--washer"></li>
-    <li class="popup__feature popup__feature--elevator"></li>
-    <li class="popup__feature popup__feature--conditioner"></li>
-  </ul>
-  <p class="popup__description">Великолепная квартира-студия в центре Токио. Подходит как туристам, так и бизнесменам. Квартира полностью укомплектована и недавно отремонтирована.</p>
-  <div class="popup__photos">
-    <img src="" class="popup__photo" width="45" height="40" alt="Фотография жилья">
-  </div>
-</article>
+    <article class="popup">
+      <img src="img/avatars/user01.png" class="popup__avatar" width="70" height="70" alt="Аватар пользователя">
+      <h3 class="popup__title">Уютное гнездышко для молодоженов</h3>
+      <p class="popup__text popup__text--address">102-0082 Tōkyō-to, Chiyoda-ku, Ichibanchō, 14−3</p>
+      <p class="popup__text popup__text--price">5200 <span>₽/ночь</span></p>
+      <h4 class="popup__type">Квартира</h4>
+      <p class="popup__text popup__text--capacity">2 комнаты для 3 гостей</p>
+      <p class="popup__text popup__text--time">Заезд после 14:00, выезд до 10:00</p>
+      <ul class="popup__features">
+        <li class="popup__feature popup__feature--wifi"></li>
+        <li class="popup__feature popup__feature--dishwasher"></li>
+        <li class="popup__feature popup__feature--parking"></li>
+        <li class="popup__feature popup__feature--washer"></li>
+        <li class="popup__feature popup__feature--elevator"></li>
+        <li class="popup__feature popup__feature--conditioner"></li>
+      </ul>
+      <p class="popup__description">Великолепная квартира-студия в центре Токио. Подходит как туристам, так и бизнесменам. Квартира полностью укомплектована и недавно отремонтирована.</p>
+      <div class="popup__photos">
+        <img src="" class="popup__photo" width="45" height="40" alt="Фотография жилья">
+      </div>
+    </article>
   </template>
 
   <!-- Сообщение об успешном создании объявления -->
