+++ js/main.js
@@ -1,5 +1,5 @@
 import './mock.js';
-import './popup.js';
+// import './popup.js';
 import './form.js';
 import './validity.js';
-// import './map.js';
+import './map.js';
